# Conflicts:
#	BookHub/BookHub/appsettings_template.json

+++ BookHub/DataAccessLayer/Entities/Order.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations.Schema;
-using BookHub.Models;
 
 namespace DataAccessLayer.Entities;
 
@@ -8,7 +7,7 @@ public class Order : BaseEntity
     public int UserId { get; set; }
     [ForeignKey("UserId")]
     public virtual User User { get; set; } = null!;
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public DateTime Date { get; set; } = DateTime.Now;
     public virtual ICollection<Book> Books { get; } = new List<Book>();
